add config-runner to docker-compose
