+++ ddd-autohaus-tactical-design-werkstattauftrag/src/main/java/ddd/autohaus/tactical/design/app/werkstattauftrag/model/auftragsposition/MaterialstatusEnum.java
@@ -1,0 +1,16 @@
+package ddd.autohaus.tactical.design.app.werkstattauftrag.model.auftragsposition;
+
+import lombok.Getter;
+import org.jmolecules.ddd.annotation.ValueObject;
+
+@ValueObject
+@Getter
+public enum MaterialstatusEnum {
+    BESTELLT("BESTELLT"), VERBUCHT("VERBUCHT"), VERFUEGBAR("VERFUEGBAR");
+
+    private String value;
+
+    MaterialstatusEnum(String value) {
+        this.value = value;
+    }
+}
