+++ ddd-autohaus-tactical-design-werkstattauftrag/src/main/java/ddd/autohaus/tactical/design/app/werkstattauftrag/model/auftragsstatus/Auftragsstatus.java
@@ -1,0 +1,8 @@
+package ddd.autohaus.tactical.design.app.werkstattauftrag.model.auftragsstatus;
+
+import lombok.Getter;
+
+@Getter
+public class Auftragsstatus {
+    private String status;
+}
