+++ ddd-autohaus-tactical-design-werkstattauftrag/src/main/java/ddd/autohaus/tactical/design/common/InvariantException.java
@@ -1,0 +1,7 @@
+package ddd.autohaus.tactical.design.common;
+
+public class InvariantException extends RuntimeException {
+    public InvariantException(String message) {
+        super(message);
+    }
+}
