--- ddd-autohaus-tactical-design-werkstattauftrag/src/main/java/ddd/autohaus/tacticald/design/WerkstattauftragApplication.java
@@ -1,13 +1,0 @@
-package ddd.autohaus.tacticald.design;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class WerkstattauftragApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(WerkstattauftragApplication.class, args);
-        System.out.println("##### WERKSTATT API STARTED #####");
-    }
-}
