--- ddd-autohaus-tactical-design-werkstattauftrag/src/main/java/ddd/autohaus/tacticald/design/app/werkstattauftrag/api/APIExceptionHandler.java
@@ -1,15 +1,0 @@
-package ddd.autohaus.tacticald.design.app.werkstattauftrag.api;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ControllerAdvice;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-
-@ControllerAdvice
-public class APIExceptionHandler {
-
-    @ExceptionHandler
-    public ResponseEntity<String> handleAnyException(Exception exception) {
-        return new ResponseEntity<>(exception.getMessage(), HttpStatus.INTERNAL_SERVER_ERROR);
-    }
-}
