--- ddd-autohaus-tactical-design-werkstattauftrag/src/main/java/ddd/autohaus/tacticald/design/app/werkstattauftrag/model/auftragsstatus/Auftragsstatus.java
@@ -1,8 +1,0 @@
-package ddd.autohaus.tacticald.design.app.werkstattauftrag.model.auftragsstatus;
-
-import lombok.Getter;
-
-@Getter
-public class Auftragsstatus {
-    private String status;
-}
