--- ddd-autohaus-tactical-design-werkstattauftrag/src/main/java/ddd/autohaus/tacticald/design/common/InvariantException.java
@@ -1,7 +1,0 @@
-package ddd.autohaus.tacticald.design.common;
-
-public class InvariantException extends RuntimeException {
-    public InvariantException(String message) {
-        super(message);
-    }
-}
