+++ ddd-autohaus-tactical-design-werkstattauftrag/src/test/java/ddd/autohaus/tactical/design/werkstattauftrag/WerkstattauftragApplicationTest.java
@@ -1,0 +1,16 @@
+package ddd.autohaus.tactical.design.werkstattauftrag;
+
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.assertTrue;
+
+public class WerkstattauftragApplicationTest
+{
+    /**
+     * Rigorous Test :-)
+     */
+    @Test
+    public void shouldAnswerWithTrue() {
+        assertTrue( true );
+    }
+}
