--- ddd-autohaus-tactical-design-werkstattauftrag/src/test/java/ddd/autohaus/tacticald/design/werkstattauftrag/WerkstattauftragApplicationTest.java
@@ -1,16 +1,0 @@
-package ddd.autohaus.tacticald.design.werkstattauftrag;
-
-import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.assertTrue;
-
-public class WerkstattauftragApplicationTest
-{
-    /**
-     * Rigorous Test :-)
-     */
-    @Test
-    public void shouldAnswerWithTrue() {
-        assertTrue( true );
-    }
-}
