--- k6-configuration-runner/docker-compose-no-api.yml
@@ -1,37 +1,0 @@
-version: '3.9'
-services:
-  otel-collector:
-    image: otel/opentelemetry-collector-contrib
-    container_name: "k6-java-otel-collector"
-    command: [ "--config=/etc/opentelemetry/otel-config.yml" ]
-    volumes:
-      - ./docker-config/opentelemetry/otel-config.yml:/etc/opentelemetry/otel-config.yml
-    ports:
-      - "4318:4318"   # OTLP HTTP
-    env_file:
-      - ./env/.env
-
-  influxdb:
-    image: influxdb:latest
-    container_name: "k6-java-influx"
-    ports:
-      - "8086:8086"
-    env_file:
-      - ./env/.env
-
-  grafana:
-    image: grafana/grafana:9.1.8
-    container_name: "k6-java-grafana"
-    volumes:
-      - ./docker-config/grafana/provisioning:/usr/share/grafana/custom/
-      - ./docker-config/grafana/my-dashboards:/usr/share/grafana/public/dashboards
-    ports:
-      - "3030:3000"
-    env_file:
-      - ./env/.env
-    environment:
-      - GF_PANELS_DISABLE_SANITIZE_HTML=TRUE
-      - GF_SECURITY_ADMIN_PASSWORD=admin
-      - GF_PATHS_PROVISIONING=/usr/share/grafana/custom/
-    depends_on:
-      - influxdb
