--- k6-configuration-runner/docker-config/opentelemetry/otel-config.yml
@@ -1,32 +1,0 @@
-receivers:
-  otlp:
-    protocols:
-      http:
-
-processors:
-  batch:
-  memory_limiter:
-    # 75% of maximum memory up to 4G
-    limit_mib: 1536
-    # 25% of limit up to 2G
-    spike_limit_mib: 512
-    check_interval: 5s
-
-exporters:
-  prometheus:
-    endpoint: "0.0.0.0:8889"
-  influxdb:
-    endpoint: "http://influxdb:8086"
-    org: "${DOCKER_INFLUXDB_INIT_ORG}"
-    bucket: "${DOCKER_INFLUXDB_INIT_BUCKET}"
-    token: "${DOCKER_INFLUXDB_INIT_ADMIN_TOKEN}"
-    metrics_schema: telegraf-prometheus-v1
-  logging:
-    loglevel: debug
-
-service:
-  pipelines:
-    metrics:
-      receivers: [otlp]
-      processors: [memory_limiter, batch]
-      exporters: [prometheus, influxdb, logging]
