--- k6-configuration-runner/docker-config/prometheus/prometheus.yml
@@ -1,8 +1,0 @@
-global:
-  scrape_interval:     5s
-  evaluation_interval: 5s
-
-scrape_configs:
-  - job_name: 'otel-collector-metrics'
-    static_configs:
-      - targets: ['otel-collector:8889'] #use service inside docker-compose as host
