+++ k6-configuration-runner/src/main/java/poc/util/HostRetriever.java
@@ -1,0 +1,15 @@
+package poc.util;
+
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.stereotype.Component;
+
+@Component
+public class HostRetriever {
+
+    @Value("${api.host}")
+    private String host;
+
+    public String getHost() {
+        return this.host;
+    }
+}
