+++ k6-configuration-runner/src/main/resources/application.properties
@@ -1,8 +1,2 @@
 server.port = 9010
-
-# available outputs: json, csv
-test.output = json
-test.breakpoint = false
-test.loops = 1
-
-path.config = poc/config/simpleConfig.json
+api.host=localhost
